requirements.dev.txt
pytest==5.0.1
pytest-aiohttp==0.3.0
